Added Injection for ALL platforms. ARM32 & Linux32 still needs testing.
Fixing focus issues so we only lose focus when you click outside the component.
Fixes a rendering issue for MacOS and Linux where symbol resolution is lazy binding and causes symbols to not be found when using RTLD_NEXT. This can cause the canvas the fallback hooks to deploy and not be as efficient as the best hooks (IE: Raw OpenGL will report a different canvas size).
Added GetModuleHandle for Linux.

Fix issues with accessing arrays as of the new API changes for accessing multiple indices at once for performance!
Proper const correctness and exception specification
Performance boost and size decrease with proper const correctness and exception handling.